src\cuda\bernoulli\cdf.cu/pmf.cu: - Fixed incorrect input type (double -> int) src\cuda\bernoulli\mgf.cu: - Fixed incorrect parameter being passed to dispatcher function (k -> t)
wrapper_utility.h:
- Consolidated repeated code in wrappers to two functions (cpu/cuda wrappers)
bernoulli, normal, and poisson wrappers:
- Consolidated repeated code into one file (wrapper_utility.h)

+++ src/cuda/bernoulli/cdf.cu
@@ -41,7 +41,7 @@ namespace fastdist::cuda::bernoulli {
 
     // Dispatcher
     void bernoulli_cdf_dispatcher(const int* k, double* output, const int n, const double p, const int stepSize) {
-        execute_cuda_kernel<double, double>(
+        execute_cuda_kernel<int, double>(
             bernoulli_cdf_kernel,
             k,
             output,
